// frontend/src/interfaces/interfaces.ts
export interface Product {
  ProductId: string;
  Name: string;
  Description: string;
  Price: number;
  OfferPrice?: number;
  OnOffer: boolean;
  Quantity: number;
  CategoryId: string;
  IsFeatured: boolean;
  IsAvailable: boolean;

  Category?: Category;
  Images?: ProductImage[];
  OrderItems?: OrderItem[];
  CartItems?: CartItem[];
  Reviews?: Review[];

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface ProductImage {
  ImageId: string;
  ProductId: string;
  ImageUrl: string;

  Product?: Product;

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface Service {
  ServiceId: string;
  Title: string;
  Description: string;
  ImageUrl?: string;
  OnOffer: boolean;
  IsFeatured: boolean;

  Requests?: ServiceRequest[];

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface ServiceRequest {
  RequestId: string;
  UserId: string;
  ServiceId: string;
  PreferredDate?: Date;
  LocationDescription?: string;
  Status: ServiceRequestStatus;

  User?: User;
  Service?: Service;

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface Category {
  CategoryId: string;
  Name: string;
  ImageUrl?: string;

  Products?: Product[];

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface User {
  UserId: string;
  FirstName: string;
  SecondName: string;
  Email: string;
  Phone: string;
  County: string;
  LocationDescription?: string;
  Role: UserRole;
  PasswordHash: string;
  IsActive: boolean;
  IsWelcomed: boolean;

  Cart?: Cart;
  Orders?: Order[];
  Payments?: Payment[];
  Reviews?: Review[];
  ServiceRequests?: ServiceRequest[];
  RefreshTokens?: RefreshToken[];
  Recoveries?: Recovery[];
  StkRequests?: StkRequest[];

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface Cart {
  CartId: string;
  UserId: string;

  User?: User;
  Items?: CartItem[];

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface CartItem {
  CartItemId: string;
  CartId: string;
  ProductId: string;
  Quantity: number;

  Cart?: Cart;
  Product?: Product;

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface Order {
  OrderId: string;
  UserId: string;
  Status: OrderStatus;
  TotalAmount: number;

  User?: User;
  Items?: OrderItem[];
  Payments?: Payment[];

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}

export interface OrderItem {
  OrderItemId: string;
  OrderId: string;
  ProductId: string;
  Quantity: number;
  PriceAtPurchase: number;

  Order?: Order;
  Product?: Product;

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}
// ...
export interface Review {
  ReviewId: string;
  UserId: string;
  ProductId: string;
  Rating: number;
  Message?: string;
  Status: ReviewStatus;

  User?: User;
  Product?: Product;

  IsDeleted: boolean;
  CreatedAt: Date;
  UpdatedAt: Date;
}